pyproject.toml
[tool.poetry]
name = "research-app"
version = "0.1.0"
description = ""
authors = ["Shinya Maeda <dosuken0122@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.339"
openai = "^1.3.3"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
unstructured = "^0.11.0"
duckduckgo-search = "^3.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
